Refactor App tests into new structure
